refactor: split handler.go into separate person and headquarters controllers
