+++ cmd/api/headquarters_controller.go
@@ -1,0 +1,1 @@
+package api
