+++ internal/domain/headquarters/errors.go
@@ -1,0 +1,1 @@
+package headquarters
