+++ internal/domain/headquarters/headquarters.go
@@ -1,0 +1,1 @@
+package headquarters
