+++ internal/domain/headquarters/service.go
@@ -1,0 +1,1 @@
+package headquarters
