+++ internal/platform/mysql/connection.go
